Fix args() test fixture
Also: factor shared fixtures out in a test_fixtures.py
